Redux files move in particular one Folder
