+++ Practical-6/src/index.tsx
@@ -2,7 +2,7 @@ import React from 'react';
 import ReactDom from 'react-dom';
 import './style.css';
 import App from './App';
-import {store} from './app/store';
+import {store} from './Redux/app/store';
 import {Provider} from 'react-redux';
 
 ReactDom.render(
